fix: e2e-v3, lint-v3
