+++ package.json
@@ -106,8 +106,8 @@
     "build:types": "yarn clean:types && yarn generate:types && yarn copy:types",
     "start:mock-server": "MOCK_SERVER_PORT=3030 nodemon --delay 500ms -e js,json --watch playground --exec 'node playground/mocks/server.js'",
     "test:parity-setup": "mkdir -p test-reports/testcafe && yarn generate-config && mockDuo=true && yarn --cwd src/v3 build --mode testcafe && yarn --cwd src/v3 serve --mode testcafe",
-    "test:parity-run": "OKTA_SIW_V3=true yarn testcafe -c 4 chrome:headless --reporter spec,xunit:build2/reports/junit/testcafe-results.xml",
-    "test:parity-ci": "yarn codegen && run-p -r test:parity-setup test:parity-run"
+    "test:parity-run": "OKTA_SIW_V3=true yarn testcafe -c6 chrome:headless --reporter spec,xunit:build2/reports/junit/testcafe-results.xml",
+    "test:parity-ci": "yarn codegen && run-p -r start:mock-server test:parity-setup test:parity-run"
   },
   "devDependencies": {
     "@babel/cli": "^7.17.6",
