--- src/v3/babel.config.json
@@ -1,16 +1,0 @@
-{
-  "presets": [
-    [
-      "@okta/odyssey-babel-preset",
-      {
-        "odyssey": {},
-        "react": {
-          "runtime": "classic"
-        }
-      }
-    ]
-  ],
-  "plugins": [
-    "@babel/plugin-transform-runtime"
-  ]
-}
