+++ src/v3/package.json
@@ -28,7 +28,7 @@
     "serve": "vite preview --port 3000",
     "test": "yarn codegen && jest",
     "test:dev": "NODE_ENV=development jest",
-    "test:e2e": "NODE_ENV=test testcafe test/e2e/tests/**/*.ts --config-file test/e2e/.testcaferc.js",
+    "test:e2e": "NODE_ENV=test testcafe test/e2e/tests/**/*.js --config-file test/e2e/.testcaferc.js",
     "test:e2e:screenshots": "yarn test:e2e -s path=test/e2e/screenshots,thumbnails=false",
     "test:report": "PREACT_APP_USE_PKCE=false NODE_ENV=test yarn jest --coverage"
   },
